+++ src/pages/Dashboard/Content/LiveKitchen/LiveKitchen.tsx
@@ -1,24 +1,24 @@
 import React from 'react';
 import {styled} from "@mui/system";
 
-const BlockDeliveredOrders = styled('div')({
-    minWidth: "255px",
+const BlockLiveKitchen = styled('div')({
+    Width: "100%",
     borderRadius: "6px",
     border: "0.5px solid #153453",
     background: "#081A2C",
     margin: "17px 0px 0px 18px",
     "@media (max-width:1900px)":{
-        minWidth: "220px",
+        minWidth: "364px",
         margin: "15px 0px 0px 14px",
     }
 });
 
-const DeliveredOrders = () => {
+const LiveKitchen = () => {
     return (
-        <BlockDeliveredOrders>
-
-        </BlockDeliveredOrders>
+        <BlockLiveKitchen>
+            
+        </BlockLiveKitchen>
     );
 };
 
-export default DeliveredOrders;
+export default LiveKitchen;
