+++ src/pages/Dashboard/Content/NewOrders/NewOrders.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import {styled} from '@mui/system';
 
 const BlockNewOrders = styled('div')({
-    minWidth: "255px",
+    Width: "100%",
     borderRadius:"6px",
     border: "0.5px solid #153453",
     background: "#081A2C",
