+++ src/pages/Dashboard/Content/SalesGraph/SalesGraph.tsx
@@ -1,25 +1,24 @@
 import React from 'react';
 import {styled} from "@mui/system";
 
-const BlockLiveKitchen = styled('div')({
-    minWidth: "360px",
+const BlockSalesGraph = styled('div')({
+    Width: "100%",
     borderRadius: "6px",
     border: "0.5px solid #153453",
     background: "#081A2C",
-    padding: "28px 22px",
     margin: "17px 0px 0px 18px",
     "@media (max-width:1900px)":{
-        minWidth: "320px",
+        minWidth: "456px",
         margin: "15px 0px 0px 14px",
     }
 });
 
-const LiveKitchen = () => {
+const SalesGraph = () => {
     return (
-        <BlockLiveKitchen>
-            
-        </BlockLiveKitchen>
+        <BlockSalesGraph>
+
+        </BlockSalesGraph>
     );
 };
 
-export default LiveKitchen;
+export default SalesGraph;
