+++ src/pages/Dashboard/Content/TableBooking/TableBooking.tsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import {styled} from "@mui/system";
 
-const BlockSalesGraph = styled('div')({
-    minWidth: "472px",
-    borderRadius: "6px",
+const BlockTableBooking = styled('div')({
+    Width: "100%",
+    borderRadius:"6px",
     border: "0.5px solid #153453",
     background: "#081A2C",
-    padding: "28px 28px",
     margin: "17px 0px 0px 18px",
     "@media (max-width:1900px)":{
         minWidth: "400px",
@@ -14,12 +13,12 @@ const BlockSalesGraph = styled('div')({
     }
 });
 
-const SalesGraph = () => {
+const TableBooking = () => {
     return (
-        <BlockSalesGraph>
+        <BlockTableBooking>
 
-        </BlockSalesGraph>
+        </BlockTableBooking>
     );
 };
 
-export default SalesGraph;
+export default TableBooking;
