+++ src/pages/Dashboard/Content/TableDetails/TableDetails.tsx
@@ -1,24 +1,24 @@
 import React from 'react';
 import {styled} from "@mui/system";
 
-const BlockProcessOrders = styled('div')({
-    minWidth: "255px",
-    borderRadius:"6px",
+const BlockTableDetails = styled('div')({
+    Width: "100%",
+    borderRadius: "6px",
     border: "0.5px solid #153453",
     background: "#081A2C",
     margin: "17px 0px 0px 18px",
     "@media (max-width:1900px)":{
-        minWidth: "220px",
+        minWidth: "275px",
         margin: "15px 0px 0px 14px",
     }
 });
 
-const ProcessOrders = () => {
+const TableDetails = () => {
     return (
-        <BlockProcessOrders>
-            
-        </BlockProcessOrders>
+        <BlockTableDetails>
+
+        </BlockTableDetails>
     );
 };
 
-export default ProcessOrders;
+export default TableDetails;
