+++ src/pages/Dashboard/Content/TableDishes/TableDishes.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {styled} from "@mui/system";
 
 const BlockTableDishes = styled('div')({
-    minWidth: "405px",
+    Width: "100%",
     borderRadius: "6px",
     border: "0.5px solid #153453",
     background: "#081A2C",
